// src/api/model/apiPlotCoordinate.ts
/*
 * Copyright(c) 2009 - 2019 Abelium d.o.o.
 * Kajuhova 90, 1000 Ljubljana, Slovenia
 * All rights reserved
 * Copyright(c) 1995 - 2018 T-Systems Multimedia Solutions GmbH
 * Riesaer Str. 5, 01129 Dresden
 * All rights reserved.
 *
 * INATrace Services API
 * Abelium INATrace Services API swagger documentation
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the openapi-typescript-angular-generator.
 * https://github.com/alenabelium/openapi-typescript-angular-generator
 * Do not edit the class manually.
 */





export interface ApiPlotCoordinate { 
    /**
     * Entity id
     */
    id?: number;
    /**
     * The latitude of the coordinate
     */
    latitude?: number;
    /**
     * The longitude of the coordinate
     */
    longitude?: number;
}

/**
 * Namespace for property- and property-value-enumerations of ApiPlotCoordinate.
 */
export namespace ApiPlotCoordinate {
    /**
     * All properties of ApiPlotCoordinate.
     */
    export enum Properties {
        /**
         * Entity id
         */
        id = 'id',
        /**
         * The latitude of the coordinate
         */
        latitude = 'latitude',
        /**
         * The longitude of the coordinate
         */
        longitude = 'longitude'
    }


    export function formMetadata() {
        return  {
            metadata: formMetadata,
            classname: 'ApiPlotCoordinate',
            vars: [
                {
                    isReadOnly: false,
                    isEnum: false,
                    required: false,
                    name: 'id',
                    classname: 'ApiPlotCoordinate',
                    dataType: 'number',
                    isPrimitiveType: true,
                    isListContainer: false,
                    complexType: ''
                },
                {
                    isReadOnly: false,
                    isEnum: false,
                    required: false,
                    name: 'latitude',
                    classname: 'ApiPlotCoordinate',
                    dataType: 'number',
                    isPrimitiveType: true,
                    isListContainer: false,
                    complexType: ''
                },
                {
                    isReadOnly: false,
                    isEnum: false,
                    required: false,
                    name: 'longitude',
                    classname: 'ApiPlotCoordinate',
                    dataType: 'number',
                    isPrimitiveType: true,
                    isListContainer: false,
                    complexType: ''
                },
            ],
            validators: {
                id: [
                ],
                latitude: [
                ],
                longitude: [
                ],
            }
        }
    }

  // export const ApiPlotCoordinateValidationScheme = {
  //     validators: [],
  //     fields: {
  //               id: {
  //                   validators: []
  //               },
  //               latitude: {
  //                   validators: []
  //               },
  //               longitude: {
  //                   validators: []
  //               },
  //     }
  // } as SimpleValidationScheme<ApiPlotCoordinate>;


}